requirements.txt
# ==========================================
# Core Framework
# ==========================================
fastapi==0.110.0
uvicorn[standard]==0.29.0

# ==========================================
# HTTP & Scraping
# ==========================================
aiohttp==3.9.3
requests==2.31.0
beautifulsoup4==4.12.3

# ==========================================
# NLP & Embeddings
# ==========================================
sentence-transformers==2.6.1
torch>=2.1.0
numpy==1.26.4
scikit-learn==1.4.1.post1

# ==========================================
# Vector Database
# ==========================================
faiss-cpu==1.7.4

# ==========================================
# Config & Validation
# ==========================================
pydantic==2.6.4
pyyaml==6.0.1

# ==========================================
# Utilities
# ==========================================
tqdm==4.66.2

# ==========================================
# Testing
# ==========================================
pytest==8.1.1
pytest-asyncio==0.23.6

# ==========================================
# Optional LLM Providers
# ==========================================
openai==1.14.3

# ==========================================
# Optional Future Enhancements
# ==========================================
# redis==5.0.3
# celery==5.3.6
# hdbscan==0.8.33
# spacy==3.7.4
# tiktoken==0.6.0